Extract query hints
